Feat: Added Hashing and salt to seed file

+++ server/api/products.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const { Cart, Product } = require('../db/models')
+const { Product } = require('../db/models')
 module.exports = router
 
 router.get('/', (req, res, next) => {
@@ -18,27 +18,33 @@ router.get('/:productId', (req, res, next) => {
 });
 
 router.post('/', (req, res, next) => {
-    Product.create(req.body)
-        .then(product => {
-            res.json(product)
-        })
-        .catch(next);
+    if (req.user.isAdmin){
+        Product.create(req.body)
+            .then(product => {
+                res.json(product)
+            })
+            .catch(next);
+    }
 });
 
 router.put('/:productId', (req, res, next) => {
-    Product.findById(req.params.productId)
-        .then(product => {
-            product.update(req.body)
-            res.send(200)
-        })
-        .catch(next)
+    if (req.user.isAdmin){
+        Product.findById(req.params.productId)
+            .then(product => {
+                product.update(req.body)
+                res.send(200)
+            })
+            .catch(next)
+    }
 })
 
 router.delete('/:productId', (req, res, next) => {
-    Product.findById(req.params.productId)
-        .then(product => {
-            product.destroy({});
-            res.send(201)
-        })
-        .catch(next);
+    if (req.user.isAdmin){
+        Product.findById(req.params.productId)
+            .then(product => {
+                product.destroy({});
+                res.send(201)
+            })
+            .catch(next);
+    }
 });
